requirements.txt
requests>=2.28
websockets>=11.0
python-dateutil>=2.8
py-clob-client>=0.34.5
